\name{alphabet.class}
\docType{data}
\alias{alphabet.class}
\alias{alphabetclass}
\title{ElectroGraph Demo Data Set}
\description{
A fictitious data set used in the ElectroGraph manual/article to
demonstrate plots and tie fidelities.
}
\usage{alphabet.class}
\format{
A 16-by-3 data frame containing sources, sinks and fidelities for a
series of ties.
}
\keyword{datasets}
